monitor: fix bytes_received and bytes_acked overflow
__u64 bytes_received;
__u64 bytes_acked;

Check struct bpf_sock_ops in
https://github.com/torvalds/linux/blob/master/tools/include/uapi/linux/bpf.h

+++ test/cmd_ss_test.go
@@ -22,7 +22,6 @@ timer:(keepalive,11sec,0) skmem:(r0,rb131072,t0,tb2626560,f0,w0,o0,bl0,d276) ts 
 ESTAB   0       0        127.0.0.1:41317    127.0.0.1:40554
 timer:(keepalive,11sec,0) skmem:(r0,rb131072,t0,tb2626560,f0,w0,o0,bl0,d276) ts sack cubic wscale:7,7 rto:206.666 rtt:6.562/12.05 ato:40 mss:32768 pmtu:65535 rcvmss:536 advmss:65483 cwnd:10 bytes_sent:312 bytes_acked:312 bytes_received:2099 segs_out:1109 segs_in:1112 data_segs_out:4 data_segs_in:4 send 399487961bps lastsnd:12722864 lastrcv:12722877 lastack:3530 pacing_rate 798975920bps delivery_rate 29127111104bps delivered:5 app_limited busy:53ms rcv_space:65483 rcv_ssthresh:65483 minrtt:0.009 snd_wnd:65536
 `
-
 	lines := strings.FieldsFunc(out, SplitNewline)
 
 	var t TcpMetric
@@ -152,11 +151,11 @@ timer:(keepalive,11sec,0) skmem:(r0,rb131072,t0,tb2626560,f0,w0,o0,bl0,d276) ts 
 	assert.Equal(uint32(1257), t.Sockets[2].BytesSent)
 	assert.Equal(uint32(312), t.Sockets[3].BytesSent)
 
-	assert.Equal(uint32(1257), t.Sockets[2].BytesAcked)
-	assert.Equal(uint32(312), t.Sockets[3].BytesAcked)
+	assert.Equal(uint64(1257), t.Sockets[2].BytesAcked)
+	assert.Equal(uint64(312), t.Sockets[3].BytesAcked)
 
-	assert.Equal(uint32(3442), t.Sockets[2].BytesReceived)
-	assert.Equal(uint32(2099), t.Sockets[3].BytesReceived)
+	assert.Equal(uint64(3442), t.Sockets[2].BytesReceived)
+	assert.Equal(uint64(2099), t.Sockets[3].BytesReceived)
 
 	assert.Equal(uint32(1112), t.Sockets[2].SegsOut)
 	assert.Equal(uint32(1109), t.Sockets[3].SegsOut)
